refactor: physics docs
