+++ src/world/mod.rs
@@ -1,10 +1,6 @@
 use bevy::prelude::*;
 pub struct WorldPlugin;
 
-//plugins
-pub mod object;
-use object::*;
-
 pub mod grid;
 use grid::*;
 
@@ -15,7 +11,6 @@ impl Plugin for WorldPlugin {
     fn build(&self, app: &mut App) {
         //systems
         //plugins
-        app.add_plugins(ObjectPlugin);
         app.add_plugins(GridPlugin);
         app.add_plugins(PhysicsPlugin);
     }
