--- src/world/object.rs
@@ -1,13 +1,0 @@
-use bevy::prelude::*;
-
-pub struct ObjectPlugin;
-
-impl Plugin for ObjectPlugin {
-    fn build(&self, _app: &mut App) {
-        
-    }
-}
-
-//----components----
-#[derive(Component)]
-pub struct Object {}
