+++ src/world/physics/mod.rs
@@ -1,0 +1,23 @@
+//! Handles the physics in the world
+
+use bevy::prelude::*;
+
+pub mod systems;
+use systems::*;
+pub mod components;
+use components::*;
+pub mod resources;
+use resources::*;
+pub mod states;
+use states::*;
+
+pub struct PhysicsPlugin;
+
+impl Plugin for PhysicsPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_systems(Update, show_colliders);
+        app.add_systems(Update, handle_colliders);
+        app.add_systems(Update, handle_smooth_movement);
+        app.add_systems(Update, handle_lerp_movement);
+    }
+}
