+++ src/world/physics/resources.rs
@@ -1,0 +1,1 @@
+use bevy::prelude::*;
